Updated to Version 1.0.1
[Changelog 1.0.1]
Minor Patch to make sure Script and application was working.
Time Variable was not being properly accessed to record the time and as
a result only 1 screenshot was being taken previously
[Upcoming]
- Optional Screen Swipe option to start at top of screen, swipe the
length of one screen, screenshot and continue until the end of the
scrollable application has been reached. (ie. useful for screenshotting
chat conversations)